+++ src/components/CellLineTable/SharedColumns.tsx
@@ -1,40 +1,18 @@
-import { Link } from "gatsby";
 import React from "react";
 import { Flex } from "antd";
 import Icon from "@ant-design/icons";
 
-import { CellLineStatus } from "../../component-queries/types";
-import { formatCellLineId } from "../../utils";
 import { WHITE } from "../../style/theme";
-import { mdBreakpoint, UnpackedCellLine } from "./types";
+import { mdBreakpoint } from "./types";
 
 const CertificateIcon = require("../../img/cert-icon.svg");
 const Tube = require("../../img/tube.svg");
 
 const {
-    cellLineId,
     actionColumn,
     actionButton,
     tubeIcon,
 } = require("../../style/table.module.css");
-
-export const cellLineIdColumn = {
-    title: "Cell Collection ID",
-    key: "cellLineId",
-    className: cellLineId,
-    dataIndex: "cellLineId",
-    fixed: "left" as const,
-    render: (cellLineId: number, record: UnpackedCellLine) => {
-        const cellLine = (
-            <h4 key={cellLineId}>{formatCellLineId(cellLineId)}</h4>
-        );
-        return record.status === CellLineStatus.DataComplete ? (
-            <Link to={record.path}>{cellLine}</Link>
-        ) : (
-            cellLine
-        );
-    },
-};
 
 export const certificateOfAnalysisColumn = {
     title: "",
